REFACTOR: Move traits for elaboration/quoting to own submodule.
