+++ packages/catlog-wasm/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod notation;
 pub mod result;
 
 pub mod model;
